add find split, check_leakage, check_data_group, get_group_info to predictor
